Refactor test helpers into new package
