+++ src/test/java/gin/test/TestRunnerTest.java
@@ -1,5 +1,7 @@
-package gin;
+package gin.test;
 
+import gin.SourceFile;
+import gin.test.TestRunner;
 import org.junit.Before;
 import org.junit.Test;
 import java.io.File;
